Add fixes for import and export
